Initial edit: new typedefs, some pointers (huehue)
Edited our structures to include the pointer to the next structure.

modified variables:
structure: pBillsType
pointer to next: pNextBill
structure: pCandidateType
pointer to previous; pPrevCand
pointer to next: pNextCand

deleted variable: arrCandidates

Most of the edits we have to make are already written on the Edmodo PDF.

IMPORTANT NOTE: the PDF says that we should "update the function displayByParty() since this function will accept a DOUBLY-LINKED LIST now." (caps mine). So...we have our work cut out for us.

I plan on working on the new bills function(s), but I'll first need a guide to creating the linked list. Not that I could probably think up a way for that myself, but...I still want to at least be able to compare Miss Nats's code. And I might end up doing it wrong anyway :/